sandbox: initial commit of game app
The Game App is an example app that explains how Contract tests could be
used for verifying the behaviour of the collaborators.

In this example the Player is relaying that Ball could be taken by only
one person. If it's already taken from someone, then he will not be able
to play with it.